reafactor: rename files in sid
