+++ latebit/sid/synth/Configuration.h
