+++ latebit/sid/synth/Note.cpp
@@ -1,4 +1,4 @@
-#include "track.h"
+#include "Note.h"
 
 #include <cstdarg>
