+++ latebit/sid/synth/Note.test.cpp
@@ -1,4 +1,4 @@
-#include "track.h"
+#include "Note.h"
 
 #include <sstream>
