+++ latebit/sid/synth/Sequencer.cpp
@@ -1,64 +1,22 @@
-#include "sequencer.h"
+#include "Sequencer.h"
 
 #include <algorithm>
 #include <memory>
+#include <stdexcept>
 
-#include "configuration.h"
-#include "oscillator.h"
-#include "track.h"
-#include "tune.h"
+#include "Configuration.h"
+#include "Envelope.h"
+#include "Note.h"
+#include "Oscillator.h"
+#include "Tune.h"
 
 using namespace std;
 
 namespace sid {
+// How many samples before the end of the note should the envelope start
+// releasing. This number is very arbitrary and based on what "sounds good".
+// Exported for testing purposes.
 const int ENVELOPE_RELEASE_SAMPLES = Configuration::getSampleRate() / 100;
-
-void Envelope::attack() {
-  state = ATTACK;
-  value = 0;
-}
-
-void Envelope::release() { state = RELEASE; }
-
-void Envelope::done() {
-  state = DONE;
-  value = 0;
-}
-
-auto Envelope::process() -> float {
-  switch (this->state) {
-    case ATTACK:
-      this->value += this->attackPerSample;
-      if (this->value >= 1) {
-        this->value = 1;
-        this->state = DECAY;
-      }
-      break;
-    case DECAY:
-      this->value -= this->decayPerSample;
-      if (this->value <= this->sustainLevel) {
-        this->value = this->sustainLevel;
-        this->state = SUSTAIN;
-      }
-      break;
-    default:
-    case SUSTAIN:
-      break;
-    case RELEASE:
-      this->value -= this->releasePerSample;
-      if (this->value <= 0) {
-        this->value = 0;
-        this->state = DONE;
-      }
-      break;
-  }
-
-  return this->value;
-}
-
-auto Envelope::getValue() const -> float { return this->value; }
-auto Envelope::getState() const -> EnvelopeState { return this->state; }
-auto Envelope::getSustainLevel() const -> float { return this->sustainLevel; }
 
 void Sequencer::setNoteForTrack(Note n, int track) {
   auto& o = oscillators.at(track);
