+++ latebit/sid/synth/Tune.h
@@ -3,7 +3,7 @@
 #include <memory>
 #include <vector>
 
-#include "track.h"
+#include "Note.h"
 
 using namespace std;
